Refactor details module navigation
- Removed old navigation graph (`DetailsNavGraph.kt`).
- Moved `DetailsApiImp.kt` to `navigation` package.
- Updated `CastDetailsRoute.kt` to use the new navigation system.
- Created `Navigator.kt` in `gallery` screen for type-safe navigation to gallery.
- Updated `CastDetailsScreen.kt` to use the new gallery navigator.
- Modified `featureApiModule.kt` to reflect the package change of `DetailsApiImp`.
