+++ app/src/main/java/com/giraffe/cineverseapp/di/featureApiModule.kt
@@ -1,7 +1,7 @@
 package com.giraffe.cineverseapp.di
 
 import com.giraffe.details.DetailsApi
-import com.giraffe.details.screens.castDetails.DetailsApiImp
+import com.giraffe.details.navigation.DetailsApiImp
 import com.giraffe.explore.ExploreApi
 import com.giraffe.explore.screen.ExploreApiImp
 import org.koin.dsl.module
