+++ presentation/details/src/main/java/com/giraffe/details/screens/gallery/Navigator.kt
@@ -1,0 +1,8 @@
+package com.giraffe.details.screens.gallery
+
+import androidx.navigation.NavController
+import com.giraffe.details.navigation.GallerRoute
+
+fun NavController.navigateToGallery(actorName: String, imageUrls: List<String?>) {
+    navigate(GallerRoute(actorName, imageUrls))
+}
